# Optimization for SNPCoverageAdapter and CRAM parsing