--- a/packages/core/util/index.ts
+++ b/packages/core/util/index.ts
@@ -494,10 +494,6 @@ export function clamp(num: number, min: number, max: number): number {
   return num
 }
 
-function roundToNearestPointOne(num: number): number {
-  return Math.round(num * 10) / 10
-}
-
 /**
  * @param bp -
  * @param region -
@@ -512,7 +508,7 @@ export function bpToPx(
   }: { start?: number; end?: number; reversed?: boolean },
   bpPerPx: number,
 ) {
-  return roundToNearestPointOne((reversed ? end - bp : bp - start) / bpPerPx)
+  return (reversed ? end - bp : bp - start) / bpPerPx
 }
 
 const oneEightyOverPi = 180.0 / Math.PI