--- a/plugins/alignments/src/CramAdapter/CramSlightlyLazyFeature.ts
+++ b/plugins/alignments/src/CramAdapter/CramSlightlyLazyFeature.ts
@@ -131,7 +131,8 @@ export default class CramSlightlyLazyFeature implements Feature {
     let sublen = 0
     if (typeof this.record.readFeatures !== 'undefined') {
       // @ts-ignore
-      this.record.readFeatures.forEach(({ code, refPos, sub, data }) => {
+      for (let i = 0; i < this.record.readFeatures.length; i++) {
+        const { code, refPos, sub, data } = this.record.readFeatures[i]
         sublen = refPos - last_pos
         seq += ref.substring(last_pos - refStart, refPos - refStart)
         last_pos = refPos
@@ -200,7 +201,7 @@ export default class CramSlightlyLazyFeature implements Feature {
           cigar += `${data}H`
           oplen = 0
         } // else q or Q
-      })
+      }
     } else {
       sublen = this.record.readLength - seq.length
     }
@@ -303,95 +304,87 @@ export default class CramSlightlyLazyFeature implements Feature {
       return []
     }
     const start = this.get('start')
-    const mismatches: Mismatch[] = []
-    readFeatures.forEach(
-      (args: {
-        code: string
-        refPos: number
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        data: any
-        pos: number
-        sub: string
-        ref: string
-      }) => {
-        const { code, pos, data, sub, ref } = args
-        const refPos = args.refPos - 1 - start
-        if (code === 'X') {
-          // substitution
-          mismatches.push({
-            start: refPos,
-            length: 1,
-            base: sub,
-            qual: qual?.[pos],
-            altbase: ref,
-            type: 'mismatch',
-          })
-        } else if (code === 'I') {
-          // insertion
-          mismatches.push({
-            start: refPos,
-            type: 'insertion',
-            base: `${data.length}`,
-            length: 0,
-          })
-        } else if (code === 'N') {
-          // reference skip
-          mismatches.push({
-            type: 'skip',
-            length: data,
-            start: refPos,
-            base: 'N',
-          })
-        } else if (code === 'S') {
-          // soft clip
-          const len = data.length
-          mismatches.push({
-            start: refPos,
-            type: 'softclip',
-            base: `S${len}`,
-            cliplen: len,
-            length: 1,
-          })
-        } else if (code === 'P') {
-          // padding
-        } else if (code === 'H') {
-          // hard clip
-          const len = data
-          mismatches.push({
-            start: refPos,
-            type: 'hardclip',
-            base: `H${len}`,
-            cliplen: len,
-            length: 1,
-          })
-        } else if (code === 'D') {
-          // deletion
-          mismatches.push({
-            type: 'deletion',
-            length: data,
-            start: refPos,
-            base: '*',
-          })
-        } else if (code === 'b') {
-          // stretch of bases
-        } else if (code === 'q') {
-          // stretch of qual scores
-        } else if (code === 'B') {
-          // a pair of [base, qual]
-        } else if (code === 'i') {
-          // single-base insertion
-          // insertion
-          mismatches.push({
-            start: refPos,
-            type: 'insertion',
-            base: data,
-            length: 1,
-          })
-        } else if (code === 'Q') {
-          // single quality value
+    const mismatches: Mismatch[] = new Array(readFeatures.length)
+    let j = 0
+    for (let i = 0; i < readFeatures.length; i++) {
+      const f = readFeatures[i]
+      const { code, pos, data, sub, ref } = f
+      const refPos = f.refPos - 1 - start
+      if (code === 'X') {
+        // substitution
+        mismatches[j++] = {
+          start: refPos,
+          length: 1,
+          base: sub,
+          qual: qual?.[pos],
+          altbase: ref,
+          type: 'mismatch',
         }
-      },
-    )
+      } else if (code === 'I') {
+        // insertion
+        mismatches[j++] = {
+          start: refPos,
+          type: 'insertion',
+          base: `${data.length}`,
+          length: 0,
+        }
+      } else if (code === 'N') {
+        // reference skip
+        mismatches[j++] = {
+          type: 'skip',
+          length: data,
+          start: refPos,
+          base: 'N',
+        }
+      } else if (code === 'S') {
+        // soft clip
+        const len = data.length
+        mismatches[j++] = {
+          start: refPos,
+          type: 'softclip',
+          base: `S${len}`,
+          cliplen: len,
+          length: 1,
+        }
+      } else if (code === 'P') {
+        // padding
+      } else if (code === 'H') {
+        // hard clip
+        const len = data
+        mismatches[j++] = {
+          start: refPos,
+          type: 'hardclip',
+          base: `H${len}`,
+          cliplen: len,
+          length: 1,
+        }
+      } else if (code === 'D') {
+        // deletion
+        mismatches[j++] = {
+          type: 'deletion',
+          length: data,
+          start: refPos,
+          base: '*',
+        }
+      } else if (code === 'b') {
+        // stretch of bases
+      } else if (code === 'q') {
+        // stretch of qual scores
+      } else if (code === 'B') {
+        // a pair of [base, qual]
+      } else if (code === 'i') {
+        // single-base insertion
+        // insertion
+        mismatches[j++] = {
+          start: refPos,
+          type: 'insertion',
+          base: data,
+          length: 1,
+        }
+      } else if (code === 'Q') {
+        // single quality value
+      }
+    }
     return mismatches
   }
 }