… plugins/alignments/package.json
… yarn.lock