This can save about ~1second in generateCoverageBins calls on a 20kb region

Allocates less nested objects


On 200x coverage shortreads from jb2profile

Time spent in generateCoverageBins specifically:

This branch 2.82s
Main branch 3.6s

Possibly we could also bin multiple bp into a single bin (jb1 does this) and or use skips_and_dels (another jb1-ism) to optimize further, but we would need to be careful about binning multiple bp into a single bin with snps/modifications as those should probably not be aggregated