@cmdcolin 2022-05-10
This PR now also adds some CRAM optimizations too (raw loops instead of foreach, preallocate mismatches). Combined with @gmod/cram optimization, this branch is about ~20% faster than main (the @gmod/cram was about 10%-15% faster) for longread CRAM

400xcram longread
main 39.545
this branch 29.794
1.2x faster

600xcram longread
main 51.299s
this branch 39.545s
1.2x faster

800xcram longread
main 63.608s
this branch 51.496s
1.2x faster