@cmdcolin 2022-05-10
note: these changes above improve the longread performance more while the @gmod/cram pr improve shortread more